farmcast-ml/pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "farmcast-ml"
version = "0.1.0"
description = "FarmCast deterministic production ML platform"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
  "fastapi>=0.110.0",
  "joblib>=1.3.2",
  "lightgbm>=4.3.0",
  "numpy>=1.26.4",
  "optuna>=3.6.1",
  "pandas>=2.2.1",
  "pyarrow>=15.0.2",
  "pillow>=10.2.0",
  "pydantic>=2.6.4",
  "pyyaml>=6.0.1",
  "scikit-learn>=1.4.1",
  "scipy>=1.12.0",
  "uvicorn>=0.27.1",
]

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-q"
